SAMU fix 22-02-2022

+++ database/migrations/2021_11_15_171463_create_samu_events_table.php
@@ -68,7 +68,7 @@ class CreateSamuEventsTable extends Migration
             $table->integer('soap')->nullable();
             $table->integer('hgt')->nullable();
             $table->integer('fill_capillary')->nullable();
-            $table->integer('t')->nullable();
+            $table->decimal('t',5,2)->nullable();
             
             $table->text('treatment')->nullable();
             $table->text('observation_sv')->nullable();
